Fix: Correct cart behavior & UI updates on Add to Cart button in Product Detail Page
Fixed the issue of the "inflecting widget" when tapping the Add to Cart button.

Updated Add to Cart button UI for better user experience when the item is added.

Disabled the button when the item is already in the cart to prevent multiple additions.

Minor UI tweaks for a smoother interface.